Remove robot blocking header
Following the current production setup guide to putting a Vue Storefront app into production will result in an app which is not indexable by search engines. This is because the linked example NGINX file contains this directive; 

`add_header X-Robots-Tag none; `

Although developers should be reviewing their NGINX config files anyway, if they are using this example file as a basis this header can be easily missed, this will result in the website being blocked by Google/Bing/Yandex.